+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Svelte changelog
 
+## Unreleased (4.0)
+
+* Minimum supported Node version is now Node 14
+
+## Unreleased (3.0)
+
+* Handle `width`/`height` attributes when spreading ([#6752](https://github.com/sveltejs/svelte/issues/6752))
+* Add support for resize observer bindings (`<div bind:contentRect|contentBoxSize|borderBoxSize|devicePixelContentBoxSize>`) ([#8022](https://github.com/sveltejs/svelte/pull/8022))
+
 ## 3.58.0
 
 * Add `bind:innerText` for `contenteditable` elements ([#3311](https://github.com/sveltejs/svelte/issues/3311))